### 2025-07-01 e4c9017
add missing DOIs
