+++ paper/paper.bib
@@ -4,6 +4,7 @@ author = {Borry, Maxime and Cordova, Bryan and Perri, Angela and Wibowo, Marsha 
 address = {United States},
 copyright = {2020 Borry et al.},
 issn = {2167-8359},
+doi = {10.7717/peerj.9001},
 journal = {PeerJ (San Francisco, CA)},
 keywords = {Microbiomes ; Paleofeces ; Anthropology ; Archaeology ; Archeology ; Bioinformatics ; Consortia ; Coprolite ; Data Mining and Machine Learning ; Datasets ; Deoxyribonucleic acid ; Diabetes ; DNA ; Dogs ; Endogenous DNA ; Feces ; Genomics ; Intestinal microflora ; Machine learning ; Mesolithic ; Metagenomics ; Microbiology ; Microbiome ; Microbiota ; Mitochondrial DNA ; Neolithic ; Sediments ; Westernization},
 language = {eng},
@@ -19,6 +20,7 @@ author = {Battillo, Jenna},
 address = {Berlin/Heidelberg},
 copyright = {Springer-Verlag GmbH Germany, part of Springer Nature 2019},
 issn = {1866-9557},
+doi = {10.1007/s12520-019-00944-y},
 journal = {Archaeological and anthropological sciences},
 keywords = {Agricultural land ; Anthropogenic factors ; Anthropology ; Archaeology ; Chemistry/Food Science ; Coprolite Research: Archaeological and Paleoenvironmental Potentials ; Corn ; Deoxyribonucleic acid ; Diet ; DNA ; Earth and Environmental Science ; Earth Sciences ; Farmers ; Farming ; Forage ; Foraging behavior ; Geography ; Life Sciences ; Nutritive value ; Original Paper ; Pollen ; Productivity ; Reliance},
 language = {eng},
@@ -35,6 +37,7 @@ author = {Wibowo, Marsha C. and Yang, Zhen and Borry, Maxime and Hübner, Alexan
 address = {London},
 copyright = {The Author(s), under exclusive licence to Springer Nature Limited 2021},
 issn = {0028-0836},
+doi = {10.1038/s41586-021-03532-0},
 journal = {Nature (London)},
 keywords = {14/28 ; 45/23 ; 631/181/27 ; 631/326/2565/2134 ; 631/326/2565/2142 ; Genomes ; Reconstruction ; Antibiotic resistance ; Antibiotics ; Archaeology ; Chronic illnesses ; COVID-19 ; Deoxyribonucleic acid ; DNA ; Evolutionary genetics ; Genetic aspects ; Humanities and Social Sciences ; Intestinal microflora ; Microbiological research ; Microbiomes ; Microbiota ; Microbiota (Symbiotic organisms) ; Microorganisms ; Mucin ; multidisciplinary ; Principal components analysis ; Science ; Science (multidisciplinary) ; Soil contamination ; Taxonomy},
 language = {eng},
@@ -51,6 +54,7 @@ author = {Taylor, Rebecca S. and Manseau, Micheline and Redquest, Bridget and Ke
 address = {Dordrecht},
 copyright = {The Author(s) 2021},
 issn = {1877-7260},
+doi = {10.1007/s12686-021-01235-2},
 journal = {Conservation genetics resources},
[…]
