### 2025-07-01 #8531 @MeriamOs
Potential reviewers with nextflow and nf-core experience: Juke34 and rcannood
