### 2025-07-02 338551d
Shortened paper
