+++ paper/paper.md
@@ -29,34 +29,30 @@ bibliography: paper.bib
 
 # Summary
 
-With the advancement of Next Generation Sequencing technologies, (palaeo)faeces have become a unique and valuable source in the fields of archaeology [@Battillo:2019], microbiome studies [@Rifkin:2020; @Wibowo:2021], and species ecology and conservation [@Ang:2020; @Taylor:2022], and even shows promise in forensic investigations [@Quaak:2017; @Quaak:2018]. To use faecal samples for such studies, the central and/or first question is often "who deposited the faeces?", before moving onto further analyses. The nf-core/coproID pipeline helps to answer this question by taking raw sequencing data and predicting the depositor's species.
+With the advancement of Next Generation Sequencing technologies, (palaeo)faeces have become a unique and valuable source in the fields of archaeology [@Battillo:2019], microbiome studies [@Rifkin:2020; @Wibowo:2021], species ecology and conservation [@Ang:2020; @Taylor:2022], and even shows promise in forensic investigations [@Quaak:2017; @Quaak:2018]. To use faecal samples for such studies, often the first question is "who deposited the faeces?", before doing additional analyses. The nf-core/coproID pipeline helps to answer this question by taking raw sequencing data and predicting the depositor's species.
 
-The raw sequencing data is first pre-processed by nf-core/coproID to trim adapters and remove low quality and low complexity reads with fastp [@Chen:2018]. Bowtie2 [@Langmead:2018] is then used to align the reads to multiple customer specified reference genomes of potential depositor (host) species. These reads are then processed with sam2lca [@Borry:2022] to retain only reads specific to one of the references, and normalised according to the size of the genome. Furthermore, a taxonomic profile is created with kraken2 [@Wood:2019], and compared to a customer supplied database of potential sources using sourcepredict to estimate the percentages of contributing sources [@Borry:2019]. Both the normalised host DNA and the sourcepredict results are used to predict the most likely depositor of the faeces. The pipeline also incorporates ancient DNA damage estimates using pyDamage [@Borry:2021] and damageprofiler [@Neukamm:2021] for authenticating the ancient nature of the host DNA, when working with palaeofaeces. All results are collated into a Quarto notebook html report for an easy overview of all the results. 
+The raw sequencing data is first pre-processed to trim adapters and remove low quality and low complexity reads with fastp [@Chen:2018]. Bowtie2 [@Langmead:2018] is then used to align the reads to multiple customer specified reference genomes of potential depositor (host) species. Next, these reads are processed with sam2lca [@Borry:2022] to retain only reads specific to one of the references, and normalised according to the size of the genome. Furthermore, a taxonomic profile is created with kraken2 [@Wood:2019], and compared to a customer supplied database of potential sources using sourcepredict to estimate the percentages of contributing sources [@Borry:2019]. Both the normalised host DNA and the sourcepredict results are used to predict the most likely depositor of the faeces. The pipeline also incorporates ancient DNA damage estimates using pyDamage [@Borry:2021] and damageprofiler [@Neukamm:2021] for authenticating the ancient nature of the host DNA, when working with palaeofaeces. All results are collated into a Quarto notebook html report for an easy overview of all the results. 
 
 # Statement of need
 
[…]
