### 2025-07-08 f6e0d6f
Template update for nf-core/tools version 3.3.2
