+++ .nf-core.yml
@@ -2,7 +2,7 @@ lint:
   files_unchanged:
     - docs/images/nf-core-coproid_logo_light.png
     - docs/images/nf-core-coproid_logo_dark.png
-nf_core_version: 3.3.1
+nf_core_version: 3.3.2
 repository_type: pipeline
 template:
   author: Maxime Borry & Meriam Van Os
