### 2025-08-04 #8531 @arfon
@MeriamOs – thanks for your submission to JOSS. Can you please point me to the software contributions of this project (i.e., novel code that you have written as authors to make this project possible?)

Also, a heads up – we're currently managing a large backlog of submissions and the editor most appropriate for your area is already rather busy.
