### 2025-08-12 #8531 @MeriamOs
@arfon thanks for the heads up regarding the queue.

Regarding my contributions: I've rewritten the pipeline in the newest Nextflow DSL2 language, and have made changes to some of the workflow steps to improve efficiency and accuracy (new workflow overview is present in the paper). I have also added reporting with quarto, which is the coproid_quarto_report.qmd file.

The changes are further briefly outlined in the CHANGELOG.md file

Changed:
- DSL2 rewrite
- fastp replaced AdapterRemoval for read quality trimming/merging
- sam2lca is now used for computing the endogenous host DNA quantity, instead of custom python scripts, allowing for handling more flexibly test host genomes
- Pipeline reporting is now performed using Quarto, instead of Jupyter notebook
- Minor JSON schema updates

New
- aDNA damage testing using PyDamage
- Interactive plots in report
- Unit testing of most modules in the pipeline
