### 2025-09-24 #8531 @MeriamOs
Hi @Kevin-Mattheus-Moerman,  just wanted to gently ask if you have any updates on how long the waitlist roughly is?
