### 2025-10-16 da41054
Template update for nf-core/tools version 3.4.1
