### 2025-10-21 #8531 @MeriamOs
> [@MeriamOs](https://github.com/MeriamOs) Following on my last comment.
>
> > [@MeriamOs](https://github.com/MeriamOs) I will be helping to edit this paper. Can you bring the Wordcount for paper.md to 1000 words or less?

@abhishektiwari I'll get on to this next week as I'm traveling at the moment sorry!
