### 2025-10-26 #8531 @MeriamOs
@abhishektiwari Just want to check, does the whole document need to be less than a 1000 words? Because without the metadata of the file, the article is 998 words.

> > [@MeriamOs](https://github.com/MeriamOs) Following on my last comment.
> > > [@MeriamOs](https://github.com/MeriamOs) I will be helping to edit this paper. Can you bring the Wordcount for paper.md to 1000 words or less?
>
> [@abhishektiwari](https://github.com/abhishektiwari) I'll get on to this next week as I'm traveling at the moment sorry!
