### 2025-10-28 #8531 @abhishektiwari
👋 @Juke34 @vinegang, would any of you be willing to review this submission for Journal of Open Source Software (JOSS)?  The submission under consideration is "nf-core/coproID v2.0: An improved pipeline for the identification of (palaeo)faecal depositors".

Thanks
Abhi
