### 2025-10-29 #8531 @abhishektiwari
Thanks

> Yes I believe so. I will just get [@Maxime](https://github.com/maxibor) (co-author) to confirm this as well.
>
> > Thanks [@MeriamOs](https://github.com/MeriamOs) - As there are [other contributors](https://github.com/nf-core/coproid/graphs/contributors) to Github project, can you confirm author list appropriate and complete?
