### 2025-10-29 #9324 @editorialbot
[REVIEW]: nf-core/coproID v2.0: An improved pipeline for the identification of (palaeo)faecal depositors
**Submitting author:** @MeriamOs (<a href="http://orcid.org/0009-0008-9835-8874">Meriam  van Os</a>)
**Repository:** https://github.com/nf-core/coproid
**Branch with paper.md** (empty if default branch): paper
**Version:** v2.0
