### 2025-11-04 nf-core/coproid#87 @johnvusich
Comments on JOSS article
Regarding [joss-reviews/issues/9324](https://github.com/openjournals/joss-reviews/issues/9324),

Using Nextflow version 25.04.4 from my institution's HPC, I ran nf-core/coproid using the test data. Here is the script that I ran:

`nextflow run nf-core/coproid -r 2.0.0 -profile test,singularity --outdir $OUTDIR -w $WORKDIR -c $CONFIG`

I got the following error that seems to be syntax related in `coproid/modules/nf-core/sam2lca/analyze/main.nf`:

```
N E X T F L O W   ~  version 25.04.4

Launching `https://github.com/nf-core/coproid` [serene_faraday] DSL2 - revision: 56308278ce [2.0.0]

ERROR ~ Module compilation error
- file : /mnt/scratch/vusichjo/nxf_assets/nf-core/coproid/modules/nf-core/sam2lca/analyze/main.nf
- cause: Variable `database` already declared in the process scope @ line 28, column 9.
       def database = database ? "${database}" : "sam2lca_db"
           ^

1 error

 -- Check script '/mnt/scratch/vusichjo/nxf_assets/nf-core/coproid/workflows/coproid.nf' at line: 10 or see '.nextflow.log' file for more details
```
