### 2025-11-12 nf-core/coproid#87 @MeriamOs
Hi @johnvusich,

Thanks for your comment. I just had a look as well. We wrote the pipeline with Nextflow version 24.10.5, which does not throw this error. I'm not sure whether I should update the pipeline to the latest Nextflow version 25.04.4 and the newest nf-core template? Or can I specify to use it only with Nextflow version 24.10.5?
