### 2025-11-13 nf-core/coproid#87 @johnvusich
Hi @MeriamOs,

I was able to get the pipeline to run using Nextflow version 24.10.2. Since the error was related to the Nextflow version, I don't think it needs to be specified within the text of the article, since the nf-core template and therefore the Nextflow version compatibility will change over time. However, please consider updating the template to be compatible with the Nextflow version ≥25.0.0.
