### 2025-11-13 nf-core/coproid#87 @johnvusich
Regarding https://github.com/openjournals/joss-reviews/issues/9324,

Please add a short section containing installation and usage instructions to the manuscript.
