### 2025-11-13 #9324 @MeriamOs
@abhishektiwari could I just check with you whether or not to include a short section containing installation and usage instructions of the pipeline to the manuscript? As asked for in https://github.com/nf-core/coproid/issues/87

Installation and usage are outlined in the README file.

I don't really see it in any other JOSS articles, so thought I would check before adding it! :)
