### 2025-11-13 nf-core/coproid#88 @johnvusich
nf-core/coproid report probability per source species
### Description of the bug

Regarding https://github.com/openjournals/joss-reviews/issues/9324,

I ran the CoproID pipeline using the test data. The pipeline ran successfully. However, when looking at the output report, there seems to be an issue with the part of the pipeline that calculates the probability of the host species.

See figure 1, probability per host species:

The figure 2 barplot from [sam2lca](https://sam2lca.readthedocs.io/en/latest/) looks as I would expect:

However, figure 3 from [sourcepredict](https://sourcepredict.readthedocs.io/en/latest/usage.html) looks like something may have gone wrong:

Here are the Kraken MultiQC plots for reference:

If I am misunderstanding the output from the pipeline, please let me know.

### Command used and terminal output

```console
module purge
module load Nextflow/24.10.2

nextflow run nf-core/coproid -r 2.0.0 -profile test,singularity --outdir $OUTDIR -w $WORKDIR -c $CONFIG
```

### Relevant files

[nextflow.log](https://github.com/user-attachments/files/23534443/nextflow.log)

[config file](https://github.com/user-attachments/files/23534453/icer.config.txt)

### System information

Nextflow version 24.10.2
nf-core/coproid 2.0.0
Hardware: HPC
Executor: Slurm
Container engine: Singularity
OS: Ubuntu 22.04.5 LTS
