### 2025-11-16 nf-core/coproid#88 @MeriamOs
@johnvusich thank you for raising this!

This was an issue with the sourcepredict reference test data rather than the pipeline. I have now updated the data, so the E. coli sample should be good now as well. Could you run the pipeline again?
