### 2025-11-18 nf-core/coproid#88 @johnvusich
@MeriamOs thank you for fixing this. The pipeline seems to work now with the test data.

Figure 1, probability per host species:

Figure 2 barplot from [sam2lca](https://sam2lca.readthedocs.io/en/latest/):

Figure 3 barplot from [sourcepredict](https://sourcepredict.readthedocs.io/en/latest/usage.html):
