### 2025-11-20 1c17fa4
Template update for nf-core/tools version 3.5.1
