### 2025-12-01 04f4cb1
Revisions JOSS review #9324
https://github.com/openjournals/joss-reviews/issues/9324#issuecomment-3576723399

The manuscript requires the following language/clarity edits:

Terminology consistency: Replace "customer" with "user" or "user-specified" throughout (lines 19, 23, 71, 83)
Corrections:
Line 64: "as followed" → "as follows"
Line 80: "Quartonotebook" → "Quarto notebook"
Line 99: "pipelines depository" → "pipeline repository"
