### 2026-02-02 #9324 @vinegang
@MeriamOs @abhishektiwari  I am looking at the paper.md at this location `https://github.com/nf-core/coproid/blob/paper/paper/paper.md`, I hope this is the latest one. Correct me if I am wrong.
