### 2026-02-02 #9324 @abhishektiwari
@vinegang yes. You can use PDF generated by EditorialBot also.
