### 2026-02-02 nf-core/coproid#87 @MeriamOs
Thanks for your comment @vinegang

I'm planning to work on updating it later this year, but if you go to the README it tells you to run it with nextflow DSL2 ≤ 24.10.0

Cheers

> Hi [@MeriamOs](https://github.com/MeriamOs)
>
> While testing the pipeline, I encountered the same compilation error( using Nextflow version 25.10.0), reported by [@johnvusich](https://github.com/johnvusich).
>
> The process in `modules/nf-core/sam2lca/analyze/main.nf` fails due to a variable name collision. In the latest versions of Nextflow (25.x), the scoping rules have become stricter. Specifically, the variable `database` is defined in the `input:`block and then re-declared using def within the `script:` block.
>
> Since the manuscript emphasizes reproducibility, updating the syntax to handle Nextflow's new scoping rules would be a big plus. It'll save users from hitting installation hurdles if they're running the most recent Nextflow releases. If a code update is not immediately possible, I would recommend adding a note to the documentation to use nextflow version <25.
