### 2026-02-04 nf-core/coproid#87 @vinegang
@MeriamOs  I did see the README, but since it currently says DSL2 ≥ 24.10.0 (which includes 25.x), I think you should fix that.
