### 2026-02-04 #9324 @vinegang
I completed my review and the submission looks good overall. The coproID pipeline is well structured and clearly implemented, with a modular nf-core design. The workflow ran successfully in testing and the approach is well suited to the biological question.

As minor revisions for improvements:

1. It would be helpful to clearly document the supported Nextflow version range (given recent 25.x scoping changes) nextflow version 24.10.3 worked for me.
2. Adding brief guidance in README for running the pipeline on HPC systems with restricted container networkfing or SSL verification will be useful. In my case, I was running it on HPC, and Singularity containers couldn't verify SSL certificates for HTTPS downloads at `sam2lca` step. I resolved this error by pre downloading the required files and bind-mounting them  into the container instead of downloading at runtime.

Overall, this is a solid and well-designed pipeline, and these documentation updates would make it even easier for users to get started across different environments.
