### 2026-02-04 83574b5
Update README.md with right nextflow badge
