### 2026-02-04 b202bc0
Update README.md with right nextflow badge
