### 2026-02-04 efaead0
Update README.md with extra info on HPC containers and working offline
