### 2026-02-05 ed35274
Update README.md small style changes
