### 2026-02-05 nf-core/coproid#87 @MeriamOs
@vinegang oops I now realise I accidentally used the wrong sign (≥ instead of ≤)! Thanks so much for picking that up! I have fixed this.
