### 2026-02-05 7376075
Update figure caption for publication
