### 2026-02-05 #9324 @MeriamOs
@abhishektiwari here the information for nf-core/coproid version 2.0.0 matching the paper title.
I have opened a PR to get the reviewers comments regarding the README incorporated for a patch release for version 2.0.1.

* [x]  Double check Paper authors and affiliations (including ORCIDs)
* [x]  Make a tagged release of your software on Github, and list the version tag of the archived version here:
                     **2.0.0**
* [ ]  Check the archival deposit (e.g., in Zenodo) has the correct title (should match the paper title)
* [ ]  Check the archival deposit (e.g., in Zenodo) has the correct authors and affiliations (should match the paper authors including order).
* [x]  Check the archival deposit (e.g., in Zenodo) has the correct version (should match the above Github tag you created).
* [ ]  Check the archival deposit (e.g., in Zenodo) has the correct ORCIDs (should match the paper authors).
* [x]  Please list the DOI of the archived version here: https://doi.org/10.5281/zenodo.15488535

@abhishektiwari The zenodo archive is automatically created when a pipeline is released with nf-core, I don't think I can make any changes to the rest of the metadata. This is the same issue as just raised at https://github.com/openjournals/joss-reviews/issues/9672
