+++ assets/multiqc_config.yml
@@ -1,7 +1,5 @@
 report_comment: >
-  This report has been generated by the <a href="https://github.com/nf-core/coproid/releases/tag/2.0.0" target="_blank">nf-core/coproid</a>
-  analysis pipeline. For information about how to interpret these results, please see the
-  <a href="https://nf-co.re/coproid/2.0.0/docs/output" target="_blank">documentation</a>.
+  This report has been generated by the <a href="https://github.com/nf-core/coproid/releases/tag/2.0.1" target="_blank">nf-core/coproid</a> analysis pipeline. For information about how to interpret these results, please see the <a href="https://nf-co.re/coproid/2.0.1/docs/output" target="_blank">documentation</a>.
 report_section_order:
   general_stats:
     order: 1000
