### 2026-02-06 717650f
Update samt2lca so pipeline executes
