+++ paper/paper.md
@@ -1,5 +1,5 @@
 ---
-title: 'nf-core/coproID v2.0.0: An improved pipeline for the identification of (palaeo)faecal depositors'
+title: 'nf-core/coproID v2.0: An improved pipeline for the identification of (palaeo)faecal depositors'
 tags:
   - Nextflow
   - nf-core
