### 2026-02-11 642cce2
Add tools update to CHANGELOG and retrieve old pipeline Chagnelog
