### 2026-02-13 #9324 @abhishektiwari
hi @MeriamOs Can you trying adding `.zenodo.json` as described below to override the metadata?

https://help.zenodo.org/docs/github/describe-software/zenodo-json/
