### 2026-02-13 #9324 @MeriamOs
@abhishektiwari if it's not a hard requirement than it's probably easier to just leave it! Otherwise I'll get in touch with the authors of nf-core/cutandrun to see how they got edit permission.

Is there anything else I need to do at this stage?

> Hi [@MeriamOs](https://github.com/MeriamOs) Ignore my last comment - If you can make changes to metadata directly on this archive then go for it. Otherwise, I got confirmation from EiC that matching title is not a hard requirement, we can work with default title generated by CI.
