### 2026-02-14 #9324 @abhishektiwari
@MeriamOs I do not see `v2.0.1` tag in target repository here https://github.com/nf-core/coproid and Zenodo archive. Can you merge your changes, tag `2.0.1` and make a new release?
