### 2026-02-14 #9324 @abhishektiwari
@MeriamOs  Please also see suggestions improvements in paper below.

Main Body (Edits in paper.md):

Line 41, 89, 91: "nf-core/coproID v2.0.0" should be "nf-core/coproID v2.0.1" assuming new tag and release.
Line 42: "Nextflow DLS2" should be "Nextflow DSL2" (Domain Specific Language 2). Note: it's correctly written as "DSL2" on line 90.
Line 47: "agains" should be "against"

References (Edits in paper.bib): Please use {} explicitly to capitalise as by default JOSS formats reference titles as sentence case.

Line 106 (Andrew, S): "Babraham bioinformatics" should be "Babraham Bioinformatics"
Line 109 (Ang et al.): "southeast asian" should be "Southeast Asian"
Line 113 (Battillo): "north american southwest" should be "North American Southwest"
Line 134 (DI Tommaso et al.): "DI Tommaso" should likely be "Di Tommaso"
Line 159 (Rifkin et al.): "mid-15th century middle iron age bantu-speaker", "Middle Iron Age" and "Bantu-speaker" should be capitalized.
Line 160: "sub-saharan african" should be "sub-Saharan African"
