### 2026-02-16 dd930ad
Fixed capitalisation in reference titles
