+++ paper/paper.bib
@@ -27,7 +27,7 @@ language = {eng},
 number = {11},
 pages = {5999-6016},
 publisher = {Springer Berlin Heidelberg},
-title = {Farmers who forage: interpreting paleofecal evidence of wild resource use by early corn farmers in the North American Southwest},
+title = {Farmers who forage: interpreting paleofecal evidence of wild resource use by early corn farmers in the {N}orth {A}merican {S}outhwest},
 volume = {11},
 year = {2019},
 }
@@ -78,7 +78,7 @@ language = {eng},
 number = {1},
 pages = {9396-9396},
 publisher = {Nature Publishing Group UK},
-title = {Faecal DNA to the rescue: Shotgun sequencing of non-invasive samples reveals two subspecies of Southeast Asian primates to be Critically Endangered species},
+title = {Faecal DNA to the rescue: Shotgun sequencing of non-invasive samples reveals two subspecies of {S}outheast {A}sian primates to be Critically Endangered species},
 volume = {10},
 year = {2020},
 }
@@ -95,7 +95,7 @@ language = {eng},
 number = {1},
 pages = {62-23},
 publisher = {BioMed Central},
-title = {Multi-proxy analyses of a mid-15th century Middle Iron Age Bantu-speaker palaeo-faecal specimen elucidates the configuration of the ‘ancestral’ sub-Saharan African intestinal microbiome},
+title = {Multi-proxy analyses of a mid-15th century {M}iddle {I}ron {A}ge {B}antu-speaker palaeo-faecal specimen elucidates the configuration of the ‘ancestral’ sub-{S}aharan {A}frican intestinal microbiome},
 volume = {8},
 year = {2020},
 }
@@ -154,7 +154,7 @@ year = {2010},
 }
 
 @article{DITommaso:2017,
-author = {{DI Tommaso}, Paolo and Chatzou, Maria and Floden, Evan W. and Barja, Pablo Prieto and Palumbo, Emilio and Notredame, Cedric},
+author = {{Di Tommaso}, Paolo and Chatzou, Maria and Floden, Evan W. and Barja, Pablo Prieto and Palumbo, Emilio and Notredame, Cedric},
 doi = {10.1038/nbt.3820},
 issn = {1546-1696},
 journal = {Nature Biotechnology 2017 35:4},
@@ -267,7 +267,7 @@ year = {2022}
 }
 
 @online{Andrews:2010,
-    title = {Babraham Bioinformatics - {FastQC} A Quality Control tool for High Throughput Sequence Data},
+    title = {Babraham {B}ioinformatics - {FastQC} A Quality Control tool for High Throughput Sequence Data},
     url = {https://www.bioinformatics.babraham.ac.uk/projects/fastqc/},
     author = {Andrews, S},
     urldate = {2023-09-20},
