+++ paper/paper.bib
@@ -102,7 +102,7 @@ year = {2020},
 
 @article{Quaak:2017,
 abstract = {The forensic science community has a growing interest in microbial population analysis, especially the microbial populations found inside and on the human body. Both their high abundance, microbes outnumber human cells by a factor 10, and their diversity, different sites of the human body harbour different microbial communities, make them an interesting tool for forensics. Faecal material is a type of trace evidence which can be found in a variety of criminal cases, but is often being ignored in forensic investigations. Deriving a human short tandem repeat (STR) profile from a faecal sample can be challenging. However, the microbial communities within faecal material can be of additional criminalistic value in linking a faecal trace to the possible donor. We present a microarray technique in which the faecal microbial community is used to differentiate between faecal samples and developed a decision model to predict the possible common origin of questioned samples. The results show that this technique may be a useful additional tool when no or only partial human STR profiles can be generated.},
-author = {Quaak, Frederike C. A. and de Graaf, Mei-Lan M. and Weterings, Rob and Kuiper, Irene},
+author = {Quaak, Frederike C A. and de Graaf, Mei-Lan M. and Weterings, Rob and Kuiper, Irene},
 address = {Berlin/Heidelberg},
 copyright = {Springer-Verlag Berlin Heidelberg 2016},
 issn = {0937-9827},
@@ -120,7 +120,7 @@ year = {2017},
 
 @article{Quaak:2018,
 abstract = {Highlights•Faecal trace material can provide valuable information in forensic investigations. •DNA extracts can be used for both human STR and microbial population profiling. •Microbial population analysis as a tool for cell type identification. •Combined report with human STR and microbial population profiling.},
-author = {Quaak, Frederike C.A and van de Wal, Yvonne and Maaskant-van Wijk, Petra A and Kuiper, Irene},
+author = {Quaak, Frederike C A. and van de Wal, Yvonne and Maaskant-van Wijk, Petra A and Kuiper, Irene},
 address = {Netherlands},
 copyright = {Elsevier B.V.},
 issn = {1872-4973},
