### 2026-02-17 #9324 @MeriamOs
@abhishektiwari I have fixed all the editorial things you raised here. :)

And here's the zenodo archive to the new patch release (2.0.1): [10.5281/zenodo.18667237](https://doi.org/10.5281/zenodo.18667237)

Note that in the paper I have changed the "2.0.0" to "2.0" to reflect the entire rewrite of the pipeline. Is this okay? There will likely be other patch releases in the (near) future. Or would you rather have it as "2.0.1" in the paper?

Also note that the issue with the later Nextflow version has been fixed now in this newer release, so it can be run with the latest Nextflow verion.

Cheers

> [@MeriamOs](https://github.com/MeriamOs) Please also see suggestions improvements in paper below.
>
> Main Body (Edits in paper.md):
>
> Line 41, 89, 91: "nf-core/coproID v2.0.0" should be "nf-core/coproID v2.0.1" assuming new tag and release. Line 42: "Nextflow DLS2" should be "Nextflow DSL2" (Domain Specific Language 2). Note: it's correctly written as "DSL2" on line 90. Line 47: "agains" should be "against"
>
> References (Edits in paper.bib): Please use {} explicitly to capitalise as by default JOSS formats reference titles as sentence case.
>
> Line 106 (Andrew, S): "Babraham bioinformatics" should be "Babraham Bioinformatics" Line 109 (Ang et al.): "southeast asian" should be "Southeast Asian" Line 113 (Battillo): "north american southwest" should be "North American Southwest" Line 134 (DI Tommaso et al.): "DI Tommaso" should likely be "Di Tommaso" Line 159 (Rifkin et al.): "mid-15th century middle iron age bantu-speaker", "Middle Iron Age" and "Bantu-speaker" should be capitalized. Line 160: "sub-saharan african" should be "sub-Saharan African"
