### 2026-02-17 #9324 @abhishektiwari
This should be fine.

> Note that in the paper I have changed the "2.0.0" to "2.0" to reflect the entire rewrite of the pipeline. Is this okay? There will likely be other patch releases in the (near) future. Or would you rather have it as "2.0.1" in the paper?
