### 2026-02-17 #9324 @abhishektiwari
@MeriamOs One more issue needs to fixed with markdown in `Output` section of paper. This is breaking pandoc PDF rendering for the list in section. Please see above PDF.

We need line break between paragraph and list.

```
The results are located in a nested folder architecture. Fourteen subfolders are created within the user identified output folder:

- bowtie2
- create
- damageprofiler
- fastp
- fastqc
- kraken
- kraken2
- multiqc
- pipeline_info
- pydamage
- quartonotebook
- sam2lca
- samtools
- sourcepredict
```
