### 2026-02-17 #9324 @abhishektiwari
@MeriamOs Please note target repository and archive are versioned as `2.0.1` without `v` prefix. This seems a convention used by nf-core repositories. Let me know if any concerns. I have updated version to `2.0.1` without `v` prefix.
