### 2026-02-17 5ed18f9
Fix output list syntax
