### 2026-02-17 7ded2a4
Fix output list syntax
