### 2026-02-17 f1e5641
Fix output list syntax
