### 2026-02-17 461ec98
Fix output list syntax
