### 2026-03-06 #9324 @rkurchin
👋 Hi, congrats on being nearly done! As AEiC for JOSS I will now help to process this submission for acceptance in JOSS. Below are some final checks, some of which may require your attention:
- [x] It is indeed merely a preference and not a hard requirement that the title of the archive match the paper – however, can you explain the difference in the author lists?
- [x] Summary, first sentence: is Next Generation Sequencing a term of art/acronym in this subfield? If not, please just change this to "next-generation sequencing"
- [x] Same sentence has broken parallelism, suggest at least changing "shows promise" to "show promise" but better would be to get rid of "shows promise" entirely and just have full parallelism to "...and even forensic investigations"
- [x] second paragraph of Summary: hyphenate "user-supplied"
- [x] end of same paragraph: capitalize HTML
- [x] in Statement of need, first sentence: missing serial comma after "health"
- [x] under Materials and Methods: I suggest including line breaks between the lettered list entries
- [x] same section: hyphenate "user-supplied"
- [x] under Workflow, number 4: hyphenate "genome-specific"
- [x] equations under Workflow: text in equations should be enclosed in `\text{}` or `\mathrm{}`
- [x] number 5: hyphenate "adapter-trimmed"
- [x] number 6: hyphenate "user-supplied"
- [x] under Output: hyphenate "user-identified"
- [x] You can entirely remove the "Discussion and conclusions" section; this is redundant and unnecessary in a JOSS paper
