### 2026-03-07 #9324 @MeriamOs
Remaining points:
As you can see, most seems in order, however the below are some points that require your attention 👇 :

* [x]  It is indeed merely a preference and not a hard requirement that the title of the archive match the paper – however, can you explain the difference in the author lists?
@rkurchin Yes, this is because of two things:
1. There were some minor contributions by others for version 1 of the pipeline, which is also in this repository, but the rewrite of the pipeline was done by Maxime and I.
2. nf-core pipelines are a community collaboration, so there have been some minor edits from other people (and the bot).

* [x]  Summary, first sentence: is Next Generation Sequencing a term of art/acronym in this subfield? If not, please just change this to "next-generation sequencing"
* [x]  Same sentence has broken parallelism, suggest at least changing "shows promise" to "show promise" but better would be to get rid of "shows promise" entirely and just have full parallelism to "...and even forensic investigations"
* [x]  second paragraph of Summary: hyphenate "user-supplied"
* [x]  end of same paragraph: capitalize HTML
* [x]  in Statement of need, first sentence: missing serial comma after "health"
* [x]  under Materials and Methods: I suggest including line breaks between the lettered list entries
* [x]  same section: hyphenate "user-supplied"
* [x]  under Workflow, number 4: hyphenate "genome-specific"
* [x]  equations under Workflow: text in equations should be enclosed in \text{} or \mathrm{}
* [x]  number 5: hyphenate "adapter-trimmed"
* [x]  number 6: hyphenate "user-supplied"
* [x]  under Output: hyphenate "user-identified"
* [x]  You can entirely remove the "Discussion and conclusions" section; this is redundant and unnecessary in a JOSS paper

@rkurchin edits as suggested all done!
