### 2026-03-07 3bbd54e
Editorial edits paper.md
