### 2026-03-15 #9324 @MeriamOs
@Kevin-Mattheus-Moerman thanks for the reply and taking over. I thought that this was preferred, but not a hard requirement as discussed with @abhishektiwari? The archive gets created automatically by nf-core and I don't have direct access to edit this

> [@MeriamOs](https://github.com/MeriamOs) I am taking over here from Rachel. I checked the points you worked on. The following point remains, please work on that:
>
> * [ ]   Please edit the archive listed title and authors to match those for the paper
